crypto-config file
Added the crypto configuration file which includes netowrk topology so that the Organization and the components under that organization can be granted certificates and keys. This is a step of building the network where the  number of peers are defined as well. I have copied the crypo-config file from the Build Your first network tutorial from hyperledger fabric.